refactor: 빙고 정보를 세팅하는 부분 코드 리팩토링
