+++ src/pages/Home/BingoSettingModal/BingoNameModal/BingoNameModal.jsx
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types';
 
 import style from './BingoNameModal.module.scss';
 import NextIcon from '../../../../Icons/NextIcon';
-import useBingoSettingStore from '../../../../stores/bingoSettingStore';
+import useBingoInfoStore from '../../../../stores/bingoInfoStore';
 
 function BingoNameModal({ setModalType }) {
-  const { bingoName, setBingoName } = useBingoSettingStore();
+  const { bingoName, setBingoName } = useBingoInfoStore();
 
   const inputChange = (event) => {
     setBingoName(event.target.value);
